# backend/docs/openapi.yaml
...
openapi: 3.0.3
info:
  title: Connect Four API (target design)
  version: 0.2.0
  description: |
    The TARGET API of the Connect Four app: what the backend already does plus what the frontend needs.

    **How this file was made.** From `frontend/docs/api-contract.md` and the TypeScript types in `frontend/src/features/*/types.ts` (the app is compiled against them),
    checked against the real backend routes. It is a DESIGN: the backend implements it step by step. The Swagger page served by the running backend (`/api/docs`,
    generated from the code) shows what exists today.

    **Status of each route** (`x-status`):
    - `implemented`: the backend already does this, exactly as written (checked with real requests).
    - `change`: exists, but must change (only `GET /matches/{id}`: spectators may read a match).
    - `new`: does not exist yet. The frontend answers it with its in-browser demo server (`x-demo-feature` names the feature switch `VITE_DEMO_FEATURES`).

    **Conventions** (the existing backend already follows these; the new routes keep them)
    - Every route is under `/api`, on the same origin as the app (an HTTPS proxy sends `/api` and `/socket.io` to the backend, everything else to the frontend).
    - Login is the httpOnly `session` cookie. Every route needs it except signup, login, 2FA verify, logout and `GET /auth/me`. Without a valid session: `401 UNAUTHORIZED`.
...
    - Ids are UUID strings. Dates are ISO 8601 in UTC.
...

...
servers:
...
  variables:
    port:
...
security:
- session: []
paths:
...
  /users:
    get:
      tags:
      - users
      operationId: searchPlayers
      summary: Search players
      description: 'The viewer is in the results too, with `friendship: self`. The Players page keeps the filters
        in its URL and sends the search 300 ms after the last key press.'
      x-status: implemented
      parameters:
      - name: search
        in: query
        required: false
        description: Part of a display name, case-insensitive
        schema:
          type: string
      - name: online
        in: query
        required: false
        description: Only players online now
        schema:
          type: boolean
      - name: friends
        in: query
        required: false
        description: Only the viewer's friends
        schema:
          type: boolean
      - name: sort
        in: query
        required: false
        description: '`name` A to Z, `wins` most wins first then by name, `newest` newest account first'
        schema:
          type: string
          enum:
          - name
          - wins
          - newest
          default: name
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                properties:
                  data:
                    $ref: '#/components/schemas/PlayerListItemPage'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /users/{displayName}:
    get:
      tags:
      - users
      operationId: getProfile
      summary: A player's profile
      description: 'The display name must match exactly, as stored (case-sensitive). The viewer''s own profile has
        `friendship: self` and `online: true`.'
      x-status: implemented
      parameters:
      - $ref: '#/components/parameters/DisplayName'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                properties:
                  data:
                    $ref: '#/components/schemas/Profile'
        '404':
          description: 'Codes: `USER_NOT_FOUND`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: USER_NOT_FOUND
                  message: See the error code table
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /users/{displayName}/matches:
    get:
      tags:
      - users
      operationId: listProfileMatches
      summary: A player's finished matches
      description: Newest first (by `endedAt`). Every logged-in user can read every player's history.
      x-status: implemented
      parameters:
      - $ref: '#/components/parameters/DisplayName'
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                properties:
                  data:
                    $ref: '#/components/schemas/ProfileMatchPage'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          description: 'Codes: `USER_NOT_FOUND`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: USER_NOT_FOUND
                  message: See the error code table
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /users/me:
    patch:
      tags:
      - users
      operationId: updateMe
      summary: Change my display name
      description: Same rules as signup. Returns the updated user.
      x-status: implemented
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: The updated user
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                properties:
                  data:
                    $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: 'Codes: `DISPLAY_NAME_TAKEN`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: DISPLAY_NAME_TAKEN
                  message: See the error code table
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /users/me/avatar:
    put:
      tags:
      - users
      operationId: uploadAvatar
      summary: Upload my profile picture
      description: 'The app crops to a square and shrinks to 256 x 256 (WebP, or PNG where WebP cannot be written)
        before sending. Check the file CONTENT, not only the type the browser declares. Serve the pictures under
        `/api` (for example `/api/avatars/<file>`): the proxy only sends `/api` and `/socket.io` to the backend.
        Give every new picture a new file name so browsers never show an old cached copy. The proxy accepts bodies
        up to 5 MB.'
      x-status: implemented
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                avatar:
                  type: string
                  format: binary
                  description: PNG, JPEG or WebP, 2 MB at most
              required:
              - avatar
      responses:
        '200':
          description: The updated user, whose `avatarUrl` shows the new picture
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                properties:
                  data:
                    $ref: '#/components/schemas/User'
        '413':
          description: 'Codes: `AVATAR_TOO_LARGE`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: AVATAR_TOO_LARGE
                  message: See the error code table
        '415':
          description: 'Codes: `AVATAR_INVALID_TYPE`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: AVATAR_INVALID_TYPE
                  message: See the error code table
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
    delete:
      tags:
      - users
      operationId: removeAvatar
      summary: Remove my profile picture
      description: 'Back to the default avatar (`avatarUrl: null`).'
      x-status: implemented
      responses:
        '200':
          description: The updated user
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                properties:
                  data:
                    $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: session
      description: 'httpOnly cookie set by signup, login and 2FA verify. The browser sends it by itself: use `credentials:
        "include"`.'
  parameters:
    Limit:
      name: limit
      in: query
      description: Page size
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 20
    Offset:
      name: offset
      in: query
      description: How many items to skip
      schema:
        type: integer
        minimum: 0
        default: 0
...
    DisplayName:
      name: displayName
      in: path
      required: true
      description: 'Matched exactly, as stored (case-sensitive). Note: nothing stops a player from choosing the
        name `me`; the routes under `/users/me` are fixed paths and do not clash.'
      schema:
        type: string
  responses:
    Unauthorized:
      description: 'No valid session: the app goes back to the login page. Code `UNAUTHORIZED`.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: UNAUTHORIZED
              message: 'No valid session: the app goes back to the login page. Code `UNAUTHORIZED`.'
    ValidationError:
      description: A body or query breaks a rule. Code `VALIDATION_ERROR`.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: VALIDATION_ERROR
              message: A body or query breaks a rule. Code `VALIDATION_ERROR`.
    RateLimited:
      description: Too many requests. Code `RATE_LIMITED`; `Retry-After` says when.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: RATE_LIMITED
              message: Too many requests. Code `RATE_LIMITED`; `Retry-After` says when.
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              $ref: '#/components/schemas/ErrorCode'
            message:
              type: string
              description: English explanation, a fallback for codes the client does not translate
          required:
          - code
          - message
      required:
      - error
    ErrorCode:
      type: string
      enum:
      - VALIDATION_ERROR
      - UNAUTHORIZED
      - FORBIDDEN
      - NOT_FOUND
      - RATE_LIMITED
      - INTERNAL_ERROR
      - INVALID_CREDENTIALS
      - EMAIL_TAKEN
      - DISPLAY_NAME_TAKEN
      - INVALID_2FA_CODE
      - TWO_FACTOR_ALREADY_ENABLED
      - TWO_FACTOR_NOT_ENABLED
      - TWO_FACTOR_NOT_PENDING
      - USER_NOT_FOUND
      - CANNOT_INVITE_SELF
      - INVALID_SETTINGS
      - TOO_MANY_WAITING_MATCHES
      - MATCH_NOT_FOUND
      - MATCH_NOT_JOINABLE
      - ALREADY_IN_MATCH
      - NOT_INVITED
      - MATCH_NOT_ACTIVE
      - NOT_YOUR_TURN
      - COLUMN_FULL
      - INVALID_COLUMN
      - BLOCKED
      - CANNOT_FRIEND_SELF
      - INVALID_INVITE
      - AVATAR_INVALID_TYPE
      - AVATAR_TOO_LARGE
      - TOURNAMENT_NOT_FOUND
      - TOURNAMENT_NOT_OPEN
      - TOURNAMENT_FULL
      - ALREADY_JOINED
      - NOT_JOINED
      - CREATOR_CANNOT_LEAVE
      - NOT_CREATOR
      - NOT_ENOUGH_PLAYERS
      description: Stable identifier of a failure. The frontend translates it (English, French, Arabic). Every code
        a route can answer is listed on the route.
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
          description: 3 to 20 letters of any language, digits, `_` or `-`; unique, case-sensitive
        avatarUrl:
          type: string
          description: 'Path under /api, for example `/api/avatars/<file>`. null: the app draws a default avatar'
          nullable: true
        locale:
          type: string
          enum:
          - en
          - fr
          - ar
        twoFactorEnabled:
          type: boolean
          description: NEW, optional. Lets the account page show only the button that applies
      required:
      - id
      - email
      - displayName
      - avatarUrl
      - locale
...
    MatchSettings:
      type: object
      properties:
        cols:
          type: integer
          minimum: 5
          maximum: 10
        rows:
          type: integer
          minimum: 5
          maximum: 9
        winLength:
          type: integer
          minimum: 3
          maximum: 5
          description: Never more than the smaller board side
        theme:
          type: string
          enum:
          - classic
          - ocean
          - sunset
          - midnight
      required:
      - cols
      - rows
      - winLength
      - theme
...
    UserSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        displayName:
          type: string
        avatarUrl:
          type: string
          nullable: true
        online:
          type: boolean
          description: Has at least one open socket (or, before sockets exist, a request in the last minute)
      required:
      - id
      - displayName
      - avatarUrl
      - online
    PlayerRef:
      type: object
      properties:
        id:
          type: string
          format: uuid
        displayName:
          type: string
      required:
      - id
      - displayName
    Friendship:
      type: string
      enum:
      - self
      - none
      - friends
      - request_sent
      - request_received
      description: How the viewer relates to another player
    ProfileStats:
      type: object
      properties:
        played:
          type: integer
        wins:
          type: integer
        losses:
          type: integer
        draws:
          type: integer
      required:
      - played
      - wins
      - losses
      - draws
      description: Finished matches only (abandoned ones do not count). played = wins + losses + draws.
    PlayerListItem:
      allOf:
      - $ref: '#/components/schemas/UserSummary'
      - type: object
        properties:
          createdAt:
            type: string
            format: date-time
            description: ISO 8601, UTC
          stats:
            $ref: '#/components/schemas/ProfileStats'
          friendship:
            $ref: '#/components/schemas/Friendship'
        required:
        - createdAt
        - stats
        - friendship
    Profile:
      allOf:
      - $ref: '#/components/schemas/UserSummary'
      - type: object
        properties:
          createdAt:
            type: string
            format: date-time
            description: ISO 8601, UTC
          lastSeenAt:
            type: string
            format: date-time
            description: When they were last online; null while online
            nullable: true
          stats:
            $ref: '#/components/schemas/ProfileStats'
          friendship:
            $ref: '#/components/schemas/Friendship'
          blocked:
            type: boolean
            description: The viewer blocked this player
        required:
        - createdAt
        - lastSeenAt
        - stats
        - friendship
        - blocked
    ProfileMatch:
      type: object
      properties:
        id:
          type: string
          format: uuid
        opponent:
          $ref: '#/components/schemas/PlayerRef'
        result:
          type: string
          enum:
          - win
          - loss
          - draw
        settings:
          $ref: '#/components/schemas/MatchSettings'
        endedAt:
          type: string
          format: date-time
          description: ISO 8601, UTC
        moveCount:
          type: integer
      required:
      - id
      - opponent
      - result
      - settings
      - endedAt
      - moveCount
      description: A finished match, seen from the profile owner's side
    UpdateProfileRequest:
      type: object
      properties:
        displayName:
          type: string
          minLength: 3
          maxLength: 20
          description: Same rules as signup
      required:
      - displayName
...
    PlayerListItemPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PlayerListItem'
        total:
          type: integer
          description: Every matching item, for the page count
        limit:
          type: integer
        offset:
          type: integer
      required:
      - items
      - total
      - limit
      - offset
    ProfileMatchPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProfileMatch'
        total:
          type: integer
          description: Every matching item, for the page count
        limit:
          type: integer
        offset:
          type: integer
      required:
      - items
      - total
      - limit
      - offset
...